Some refactoring work.
Remoed alot of code before starting to add more. Trimmed as much as
possible to get to the point that I understood what was going on in each
place.

Split the server code into multiple objects to help my sanity. Arbiter
is now the main class in the master process that keeps track of child
processes and so on and such forth.

The Worker class is responsible for handling incoming requests from the
server socket passed to its constructor.