Apply patch from Chris Lamb <lamby@debian.org> with minor revision.
Validate models before allowing connections via gunicorn_django

This prevents issues where the site is accepting connections but not all
of the models have loaded yet. If your model importing has side effects
(monkey-patching, etc) this can results in errors about missing attributes
or features simply because the code that enables those features as not been
run yet.

This issue does not affect the "run_gunicorn" management command as that
performs it's own model validation before allowing connections, so we are
simply making this consistent here.

Signed-off-by: Chris Lamb <lamby@debian.org>