Don't call reap_workers() from the SIGCHLD handler.
This fixes a bug that could be evinced as follows:

    (1) murder_workers() notices a timed out worker, sends it SIGKILL.
    (2) murder_workers() returns, main run() loop continues.
    (3) Nothing much happens, murder_workers() is called again.
    (4) While iterating over WORKERS, SIGCHLD arrives and causes
        reap_workers() to be called, which itself calls
        worker.tmp.close().
    (5) Control returns to murder_workers(), which calls
        worker.tmp.last_update(), which blows up because tmp is closed.

It looks like someone tried to paper over this problem by catching
the ValueError that is raised.  Unfortunately, ValueErrors are not the
only possible error; I have also seen OSError(9) thrown, which resulted
in gunicorn dying completely.

My fix, removing the reap_workers() call from the SIGCHLD handler,
should be safe, because reap_workers() will be called upon the next
iteration of the main run() loop anyway.