fix django example application (again)
With this change, ./manage run_gunicorn works perfectly. However, due
to oustanding issues being discussed about PYTHONPATH and import magic,
running `gunicorn_django djangotest" results in the `testing` module
failing to load. This discussion really comes down to whether or not
gunicorn should make any attempt to modify sys.path. Of note is that
when invoked as `gunicorn_django` os.getcwd() is not in sys.path.