Fix using WSGI handlers with tornado worker
Without this, tornado.web.Application instances could be used as
a gunicorn application, but running a standard wsgi callback under
the tornado worker would fail.

``self.wsgi`` is the property that may be a tornado.web.Application.
The transformed application, whether originally tornado.web.Application
or a WSGI callable, should be the argument to HTTPServer.

Also, it's probably poor form to be mutating the properties of ``self``
here when a local variable suffices.