Attempt to get stack trace from a timed out worker
This changes the arbiter to send SIGABRT to the worker one second
before it sends SIGKILL, giving the new handle_sigabrt handler time to
generate a clean stack trace of where the worker is blocked and log it.

Existing behaviour is maintained if the worker does not die, however the
log message has been changed to reflect that a stack trace was not
obtained.