Update install.rst
I noticed on the website that the documentation regarding Debian was aging, so I took the liberty of adding some new content and making modifications to the source here. I used the existing content as a template to expand the Debian installation section to account for the recent changes.

Debian "wheezy" was released in May 2013 moving "squeeze" to "oldstable" status.

I hope this helps.