bump to 17.5
Not: With this release, the versionning of Gunicorn is changing.
Gunicorn is stable since a long time and there is no point to release a
"1.0" now. It should have been done since a long time. Also from the
beginning we have only 2 kind of releases:

- major release: releases with major changes or huge features added
- interval releases: fixes and minor features added

So from now we will apply the following versionning

    R<major>.<interval>

for example R17.5 means we are relasing an interval release on the 17th
major version.