Forward-compatibility for gevent graceful shutdown
This commit closes the loop on fd6c712.

Aforementioned commit fixed observable problems in my testing with
gevent 1.0 release candidates but the explanation given at the time
was curiously bogus. It came from a misreading of pools and servers
as used in the ggevent worker.

With this change, both versions of gevent should support graceful
and non-graceful shutdown.