initial step to use shared memory
Instead of opening a fd (file or IPC) to notify the master that the worker is still
alive we could just share the memory between the process and the master
and read when it has been updated by reading a value. It will improve
the performances and use less fds.

For now let just start with a naive implementation, the mapped memory
isn't freed yet and we don't reuse the available blocks. The last point
is not important for now since we are only sharing one variable between
the worker and the master for now but it may change soon. Coming is a
malloc to reuse the blocks and allows us to share more variables between
processes for future features.

This iimplementation is also compatible with platforms not supporting
ctypes like platforms using selinux with execmem denial. A ffallbaclk to
the current notification system will be provided for platforms not
supporting mmap or ctypes.