Improved the documentation that is packaged in the RPM.
Until now, only the documentation in source form was packaged.
This changeset improves the RPM generator so that the compiled HTML
version of the documentation is packaged. Also, other informative
documents (README, LICENSE, NOTICE, THANKS) and an example server
configuration file have been included.

Finally, the Sphinx documentation's conf.py has been improved so
that it is easier to compile the docs using the default theme.