Change --inactive-master to --active-master
The reason for this change is that the default will be changed in
the future to inactive master so the option (without any argument)
should be to make it active.

For the next release the default has been set to 'True' so that we
have time to test that there are no issues with stuck workers in the
wild.

I also did some PEP8 cleanup and readibility improvements.