Fix microseconds logging, add decimal seconds.
Apache will log %D microseconds over 1 million for requests that take over a second, this change sync's that access log behavior since the formatting seems to be based on that. So a request that takes 1 second and 30 milliseconds will have 1000030 for this atom.

This pull request  also adds a %L atom for decimal seconds for a quick human-readable latency number. In the above example it would output 1.000030 to the log.