Handle SIG_CHLD via sig_queue like other signals.
The new `DYING_WORKERS` list means that we don't need to run
`reap_workers` immediately on SIG_CHLD; and in fact, we shouldn't.

The original code calling `reap_workers()` directly in a signal handler
at any time was likely unsafe to begin with: `reap_workers` calls
`os.waitpid()`, calls `self.cfg.worker_exit()`, and modifies
`self.WORKERS`. All of those are potentially unsafe in a signal handler.
And not sure what raising It also can raise `HaltServer`, which wouldn't
be caught and turned into `self.halt` in the main loop as intended.