Increase sig_queue max to 256.
After going through git blame, there was no rationale recorded for the
very low limit: it was present in the first commit introducing the
signal queue. Since SIG_CHLD gets queued now, it would be entirely
possible for the previous limit of 5 queued signals to be exceeded
during `manage_workers`.