Respect defaults when running a paster application
The values of both the 'workers' and 'umask' settings have global
defaults. In the case of 'workers', this global default can be
overridden by the WEB_CONCURRENCY environment variable, which prior to
this commit is ignored by `gunicorn --paste`.

Removing this duplication of defaults allows `gunicorn --paste` to
respect settings from:

- the environment (WEB_CONCURRENCY)
- the paste configuration file
- CLI flags