Watchdog based reloader
Replace the reloader with a reloader based on the watchdog package.
This new reloader should work on all the platforms. Most popular
platforms use an underlying event system like inotify or kqueue, but
watchdog will fall back to a polling implementation.

Start the reloader after loading the WSGI entry point of the application
so that the reloader picks up the application code modules when scanning
sys.modules and have the worker send SIGQUIT to itself in its callback
because that seems to make exits happen more quickly, which is more in
line with the common development use case for reloading.