Discard posted data if it's not consumed by application
If we receive POST but WSGI application does not consume all posted data from
the stream, we just close the socket, resulting in an abortive session close
with TCP RST without graceful finish.

This was observed when HTTP POST was sent via XHR over HTTPS.  If this happens,
Chrome displays ERR_CONNECTION_RESET and we cannot see response payload.

The proposed fix is to discard data in the body before closing the socket as
Eventlet does in `HttpProtocol.handle_one_request`.