Provide guidance on X-Forwarded-For access log.
The doc change introduced in #1037 is initially helpful but then internally
inconsistent. It correctly points out that X-Forwarded-For is no longer
used in setting REMOTE_ADDR (c4873681299212d6082cd9902740eef18c2f14f1), but
then confusingly indicates a solution using the X-Forwarded-For header.

The deployment doc provides a full configuration example which includes
proxy headers as set by Nginx. What is missing, before this patch, is a
clear suggestion on how to make use of the header in gunicorn's access
log.

Accordingly, remove the confusing suggestion and replace it with a drop-in
replacement for gunicorn's default access log format.