Adjust literalinclude in custom app doc
The goal is to include the code for the standalone_app.py, except for
the comment section at the start. The file became shorter after edits,
so the directive "lines 11-60" was returning a warning
"line number spec is out of range(1-55)".

Instead, use start-after, to specify the last comment line, and
lines: 2-, to skip the first blank line. This should be more flexible
for future changes to the example.