Reap workers in the main loop
Handle SIGCHLD like every other signal, waking up the arbiter and
handling the signal in the main loop rather than in the signal handler.
Take special care to reinstall the signal handler since Python may not.

Clean up workers and call the worker_exit hook in only one place. When
killing a worker, do not clean it up. The arbiter will now clean up the
worker and invoke the hook when it reaps the worker.

Ensure that all workers have their temporary watchdog files closed and
that the arbiter does not exit or log about other child processes dying.

With reaping handled in the main loop and kill_worker delegating
responsibility for cleanup to the reaping loop, iterate over the workers
dictionary everywhere else without concern for concurrent modification.