Move init/wrap_socket to handle phase
Proposed to move the ssl wrap_socket to the handle phase so a
misbehaving SSL client only blocks a single thread.
Could happen when handshake is not happening on time or at all
and the client keeps the connection open. Multiple misbehaving
clients can still cause thread exhaustion through blocking and
so a worker timeout.
The sync worker is subject to the same problem
(causing a worker timeout), but there is no simple
improvement for that.