Add boot_timeout setting
Summary: We have an issue with our use of gunicorn where the workers
frequently timeout during boot due to our 29s timeout. That is a
reasonable timeout once they start handling requests, but it doesn't
really make sense for booting, which is slower and not bounded by
Heroku's 30 second max response time for requests.

Unfortunately, gunicorn currently only lets us set a single timeout that
controls both the worker booting and responding to requests. This adds
a new setting to guicorn, `--boot-timeout`, which would allow us to set
a longer timeout for workers when they are booting.

To preserve backwards compatability with existing gunicorn
configurations, I have boot-timeout default to 0. One difference between
boot-timeout and timeout is that I have the arbiter interpret a
boot-timeout value of 0 as meaning it should fallback to the timeout
setting. This should mean this can be committed without changing any
existing behavior, though it does mean there is a difference between
what a value of 0 means for timeout vs boot-timeout. Let me know what
you think the right trade-off here is.

I tried to adhere to the contribution guidelines here. I have not added
unit tests, but can look into that if this seems like a change you'd
want to merge. I also ran the command to update `settings.rst` but it
seemed to generate other changes as well to things I hadn't touched, so
I left that out for now. I did run the existing test suite and confirmed
they passed after this change.

Relevant Issue: #2196

Test Plan: Added print statements to `murder_workers` to print whether
or not the worker was considered booted, and what timeout was applied.
Used that along with a print statement in `post_worker_init` to confirm
that the worker was correctly considered booted after it was
initialized.

First, tried a config where `timeout` was set to a non-zero value but
`boot-timeout` was 0 (the default) and confirmed that we use the
`timeout` value before and after the worker was booted.

Next, tried a config where `timeout` was set to a non-zero value and
`boot-timeout` was also set to a non-zero value. Confirmed we used the
`boot-timeout` value during worker initialization, and `timeout`
afterwards.

Finally, tried a config where `timeout` was set to 0 and
`boot-timeout` was set to a non-zero value. Confirmed we used the
`boot-timeout` value during worker initialization, and did not apply a
timeout afterwards.