De-duplicate and parameterize exit error logging.
The signal log message had the PID and signal in the message format string,
which makes each message unique. This is awkward in log processing with for
example Sentry, where you want to see similar messages grouped together.

The exit message was emitted twice, from the "if exitcode != 0" and
the "if exitcode > 0" clause. You would be seeing things like:

[2023-08-17 09:50:10 +0200] [66450] [ERROR] Worker (pid:66451) exited with code 1
[2023-08-17 09:50:10 +0200] [66450] [ERROR] Worker (pid:66451) exited with code 1.