Notify workers of quick shutdown after graceful shutdown times out.
As per https://github.com/benoitc/gunicorn/issues/3385 this allows to
distinguish between graceful and abnormal termination and gives a chance to
workers to do something about it (e.g. to log a stack trace).

For backward compatibility, we keep the old 2 signals model if
quick_shutdown_timeout is set to 0.